Split PR validation from protected DOT execution
