--- .github/workflows/split-dot-query-workflow-v1.yml
@@ -1,39 +1,0 @@
-name: Stage secure DOT query workflow split
-
-on:
-  push:
-    branches: [science/dot-query-selective-r11-r30-v1]
-    paths:
-      - ".github/workflows/split-dot-query-workflow-v1.yml"
-      - "tools/split_dot_query_workflow.py"
-
-permissions:
-  contents: write
-
-concurrency:
-  group: split-dot-query-workflow-v1
-  cancel-in-progress: false
-
-jobs:
-  stage:
-    if: >-
-      github.repository == 'IPS-Stuttgart/Prob4D' &&
-      github.actor == 'FlorianPfaff' &&
-      github.ref == 'refs/heads/science/dot-query-selective-r11-r30-v1'
-    runs-on: ubuntu-latest
-    timeout-minutes: 15
-    steps:
-      - name: Check out exact split revision
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-        with:
-          ref: ${{ github.sha }}
-          fetch-depth: 2
-          persist-credentials: true
-          clean: true
-      - name: Stage ordinary-file workflow payloads
-        shell: bash
-        run: |
-          set -euo pipefail
-          test "$(git rev-parse HEAD)" = "$GITHUB_SHA"
-          test -z "$(git status --porcelain=v1 --untracked-files=all)"
-          python3 tools/split_dot_query_workflow.py
